requirements.txt
django==5.0.6
djangorestframework==3.15.1
markdown==3.4.1